feat: stream-activity-feed function
