+++ env.js
@@ -5,6 +5,7 @@ const env = {
   "EntityTable": `portals-me-${process.argv[2]}-entities`,
   "SortIndex": `DataTable`,
   "FeedTable": `portals-me-${process.argv[2]}-feeds`,
+  "EntityStreamFanoutTopicArn": "arn:aws:sns:ap-northeast-1:941528793676:portals-me-dev-entity-stream-hook",
   "IdentityPoolId": process.env.COGNITO_IDP || fs.readFileSync('./token/cognito.key', 'utf8'),
   "JwtPrivate": process.env.JWT_PRIVATE ? process.env.JWT_PRIVATE.split('\\n').join('\n') : fs.readFileSync('./token/jwtES256.key', 'utf8'),
   "JwtPublic": process.env.JWT_PUBLIC ? process.env.JWT_PUBLIC.split('\\n').join('\n') : fs.readFileSync('./token/jwtES256.key.pub', 'utf8'),
