+++ functions/stream-activity-feed/lib/lib.go
